+++ packages/ub-pub/ub-pub.js
@@ -547,7 +547,9 @@ function ubGlobalErrorHandler (msg, file, line, column, errorObj) {
       detail += 'file: "' + file + '" line: ' + line
     }
     let strace = errorObj.stack || ''
-    detail += strace.replace(/\?ubver=\w*/g, '').replace(/\?ver=\w*/g, '') // remove any versions
+    detail += strace
+      .replace(/\?ubver=\w*/g, '').replace(/\?ver=\w*/g, '') // remove any versions
+      .replace(/\\n(?!\d)/g, '\n\t') // beatify stack trace
     detail = detail.replace(new RegExp(window.location.origin.replace(/:/g, '\\$&'), 'g'), '') // remove address if same as origin
     detail = detail.replace(/\/[\w-]+\.js:/g, '<b>$&</b>&nbsp;line ') // file name is BOLD
     detail = detail.replace(/\n/g, '<br>&nbsp;&nbsp;')
@@ -563,7 +565,8 @@ function ubGlobalErrorHandler (msg, file, line, column, errorObj) {
     message = errorObj && (typeof errorObj === 'string') ? errorObj : msg
   }
   if (errorObj && errorObj.detail) {
-    detail = errorObj.detail + (detail ? '<br/>' + detail : '')
+    detail = errorObj.detail.replace(/\\n(?!\d)/g, '\n\t\t') + // beatify stack trace
+      (detail ? '<br/>' + detail : '')
     // 405 Method Not Allowed
     if (errorObj.detail === 'Method Not Allowed') {
       message = 'recordNotExistsOrDontHaveRights'